Update list of supported infix operators
Division has been removed because `(/)` does not produce
an `Int` in Haskell. Functions like `div`, `mod` or `quot`
and `rem` should be added as part of a `Prelude` module.
This does also apply to functions like `not`.

The power operator `(^)` has been added to the list of
supported arithmetic operatos.