Add Coq base library #5
The library contains the definitions of `Free` and `Container` as well
as the functors `Zero` and `One` and their container instances as
presented in 'One Monad to Prove Them All'. There is also the `Partial`
type class described in `notes/free-transformation` and all relevant
predefined types and operations.

This library makes it easier to write example code. To use the Free
monad or one of the type classes write

    From Base Require Import Free.

The functors and their instances are defined in
`Free.Container.<Functor>` and need to be included separately.

The predefined data types and operations can be used by adding the
following command:

    From Base Require Import Prelude.