Add custom source span for messages #18
The `haskell-src-exts` source spans do not provide
the source code that contains the source span.
To pretty print reported messages we need the
line of code that corresponds to a source span.
The new data type contains this information.

The parser module was updated to automatically
convert the source spans generated by `haskell-src-exts`
to the new format such that we do not have to perform
a conversion anywhere else.