Unify translation of build-in and defined types #5
Previously we had to distinguish whether a type was pre
defined or defined by the user because only in the latter
case we added the parameters `Shape` and `Pos`. Now we
also add those parameters to `Int`, `Bool` and `Unit`
such that we do not have to handle these special cases.