Fix mutually recursive data types #30
All identifiers of data type declarations in the same strongly
connected component need to be added to the environment before
any declaration is converted, such that they can depend on each
other. We also declare constructors beforehand such that when
there is a constructor with the same name as a type variable
(after conversion to lower case), renaming the type variable
is prefered over renaming the constructor.