Store arities in environment #33
To perform the eta conversion when converting partial function
or constructor applications we need to know the number of arguments
expected by functions and constructors. For good measure we also
store the number of type arguments expected by type constructors.
This information may be used in the future to warn the user e.g.
when a type constructor is insufficiently applied.