Add module for extraction of dependencies #34
In the implementation of depenency graphs we already have
functions that extract the names of (type) variables and
constructors a (type) expression depends on. To implement
the translation of function declarations we need to find
out the type variables used in its type signature. We can
reuse the code for constructing dependency graphs here.

The code has been moved to a new module that will be
shared among the dependency graph module and the converter.
Utility functions have been added to find specific types
of dependencies (e.g. only type variables not type constructors).