Simplify lambda abstractions on right hand sides #40
If there is a lambda abstraction on the right hand side of a
function declaration, we can eliminate the lambda abstraction
by adding the arguments to the function declaration.

This step allows the translation of recursive functions whose
outermost expression is a lambda abstraction. Nested lambda
abstractions are still not allowed in recursive functions.