Fix spell check script #55
There was a bug in the regular expression that removed minted
code blocks that prevented the script from finding misspelled
words between the first and last code block of a file.
We now use `perl` to match regular expressions instead of `sed`
because the latter does not support non-greedy quantifiers.