Add workaround for `\toCoq` command #55
The `\toCoq` command did not work properly in some situation.
This was the case because minted inserted a space after the
command name (i.e. before the argument). Because minted also
changes the catcode of spaces, the inserted space was not
ignored by LaTeX.
We can work around this issue by defining a new command, that
handles the cat 13 space after the command name explicitly.