Don't annotate the return type of helper functions #54
In general we cannot know the return type of a helper function
because we perform no type inference.
This also explains why I had to add a special case for lambda
abstractions in outer most positions.
I keep the workaround because it still leads to the generation of
more readable code.