Extend spell checker to warn about missing commas #60
As I commonly forget commas after introductory words / phrases
I extended the spell checker script to look for words listed
in https://www.smart-words.org/linking-words/transition-words.html
If there is any such word, that is not followed by a comma before the
end of the sentence, a warning is produced.
This script is fairly rudementary and cannot detect missing commas
in all situations (or might produce false alarms).