Annotate return type before inlining #15
Due to the lack of type inference, the return type of helper
functions is not known by the compiler. There are some cases
where Coq cannot infer their return type either (see #11).
We can help Coq to infer the types of helper functions by
annotating the recursive calls with the return type provided
by the user.