Improve performance of unit tests
Previously the Prelude module interface file had to be loaded
in every test case. This had a significant impact on the test
performance if converters were used in QuickCheck properties.
Now we are using a global `IORef` and `unsafePerformIO` to
buffer the loaded Prelude module interface. Therefore, the
file has to be loaded only once.