Always generate `From ... Require ...` sentences for imports
During a previous merge commit the generation of QuickCheck
imports was accidentally removed. This is now fixed.

Additionally imports of all generated modules now use a
`From Generated Require Export` sentence. The logical path
`Generated` is automatically added to the `_CoqProject` file
and refers to the output directory. This fixes an issue where
the `.v` files would compile using the Makefile but not inside
the CoqIDE.