Avoid rebuilding dependencies in GitHub actions
When Cabal's package list is updated, the dependencies need to be
rebuild. However, the updated dependencies are not cached because
the Cabal files in the repository did not change (and therefore,
there is a direct cache hit which prevents a cache update).
The cache keys now contain the hash of Cabal's index file.
This prevents a direct cache hit and causes the updated dependencies
to be cached. As a result subsequent jobs and pushes don't have
to rebuild the dependencies again.