Add visible type applications after type inference
Before we added type applications with fresh type variables first
and inferred the actual type of the type arguments later.
However, in case of recursive functions, it is still necessary
to add type arguments in a post processing step because the type
arguments are not known prior to type inference.
Thus, we add all type arguments after type inference now.
Adding type arguments to recursive calls still has to be implemented.