Use record syntax for all AST nodes
We are already using record syntax for big AST nodes such as
modules and function declarations. For the sake of consistency,
we are now using record syntax everywhere.
This eliminates the need for 'GetSrcSpan' and 'GetDeclIdent'
type classes. Additional selectors were added to improve the
readability of the code. It might be usefull to add more selectors
in the future.