Add optional type annotation to every AST node
It is currently not possible for us to annotate the type of
recursive helper functions since the type of `case` expressions
is not known after type inference. Thus, the idea is to annotate
the type of every sub-expression during type inference.
This could also simplify the AST by removing the need for
an expression type signature node.
We still have to update the code to use (or ignore) the new
field of the constructors of `Expr`.