Remove expression type signature nodes from the AST
The addition of expression type annotations in 08080a9 makes
the `ExprTypeSig` constructor redudant. The new approach cannot
handle multiple type signatures for the same expression, though.
Therefore, expressions of the form `(e :: τ₁) :: τ₂` are not
supported anymore. A warning is reported in those cases and the
type signature for `τ₂` is ignored. In theory it would be possible
to (automatically) translate `(e :: τ₁) :: τ₂` to something like
`id (e :: τ₁) :: τ₂` to avoid this restriction. However, I don't
think that this pattern is common enought to justify this extra
step.