Store entries by their original names in environment
The resolver pass converts all references in the AST to their
original names. Therefore, we only have to lookup original names
in the environment. We don't have to worry about abigous names
and nesting levels. When importing module we don't have to make
a distinction between hidden and exported entries anymore. The
resolver pass rejects references to hidden entries already.