Remove optimization for functions with lambda on right-hand side
Funcstion declarations of the form

    f x₁ … xₙ = \y₁ … yₘ -> e

were transformed to

   f x₁ … xₙ y₁ … yₘ = e

before, such that the generated Coq code is more readable.
However, while the following function declaration is valid

    f x = \x -> x

since arguments of lambda functions can shadow other local variables,
the transformed version is not

    f x x = x

since there can be no duplicate argument declarations.
Thus, the transformation is not allowed and had to be removed.