Remove workaround from resolver pass
The workaround added in 7e19af4 to fix the tests has been removed
in favour of another workaround in the testing library.
Instead of adding all entries from the environment to the resolvers
internal state, test entries are now added to their own module
which is explicitly imported by the dummy modules created during
testing.