Don't use `Converter` monad in substitutions
This `Converter` monad was used in substitutions to generate
fresh identifiers for variable patterns and type arguments.
This approach lead renamed type variables even if they did
not have to be renamed which reduced the readablility of the
generated code. The new approach does rename variables only
if there would be a name conflict otherwise.