Force recompile in `full-test.sh`
When there are warnings and the user runs `test.sh` (which
just reports the warnings but does not fail) and `full-test.sh`
would not recompile all modules of the compiler, it would not
spot the warnings. The CI pipeline does not have the same issue
since it never invokes the debugging scripts.