Implement generation of Coq properties from QuickCheck computations
With this change it is no longer necessary for the compiler
to generate `Theorem` sentences and for the user to copy the
generated theorems to their own `.v` file in order to prove
the QuickCheck properties. Instead QuickCheck properties are
treated as regular functions which compute a QuickCheck property
which can be converter to a Coq property using the `quickCheck`
function from the `Test.QuickCheck` module in the base library.