Add custom equality module renaming for IR AST
In tests we often have to compare the results of transformations
on the AST with the expected outcome. Before we have used pretty
printing combined with ignoring whitespace to test for eqaulity
(since we could not create an AST easily). Thanks to the new IR
parser, it will be feasible to test for equality on the nodes
directly. However, that does not work in the presence of source
spans and internal identifiers (since the whole idea of internal
identifiers is that they cannot be parsed directly).
The new similarity test for IR ASTs solves both problems by
ignoring source spans and comparing the AST such that bound
variables with different names are considered equal if they
are bound by the same binder.