Require explicit forall in test environment entry types
Since in Haskell type variables are implicitly quantified but in
IR they must be quantified explicitly, it is easy to forget to
quantify type variables when writing the tests. If there are free
type variables in the type schema for an environment entry and
no explicit forall, an error is printed such that the tester can
spot the potential mistake.