Move return type field of function declaration node
In the parser, the return type is on the left-hand side, but in the
AST it still was listed after the right-hand side. Since free type
variables are ordered from left to right it has not been clear whether
type arguments from the retun type preceede the type arguments from
the right-hand side or not. Now the AST is consistent with the parser.