Test running CI pipeline in pull requests (#1)
* Run CI pipeline on MacOS and Windows in addition to Ubuntu

* Use fork of `setup-haskell` action that supports Windows and MacOS

* Disable Windows in CI pipeline

Caching does not work on windows at the moment, because the
paths are different on Windows.

* Specify Cabal version more precisely in CI pipeline

Under MacOS we have to specify the full version.

* Fix CI pipeline for MacOS

* Test examples on MacOS

* Skip deployment of documentation if `DEPLOY_KEY` is not set

* Enable CI pipeline in pull requests

* Disable MacOS in CI pipeline

Eventhough the CI pipeline works on MacOS, the run time costs
of installing GHC are too high at the moment. Every minute on
MacOS is multiplied by 10. If each step involving MacOS takes
five Minutes or more we will quickly exceed our time limit.
Until Haskell is officially supported for MacOS on GitHub
Actions, we should continue to test the compiler on Linux only.

Windows has the same problem that GHC needs to be installed.
However, the minute multiplier is just 2 in this case.
There is the additional problem that the shell scripts would
have to be updated, though. Thus, the compiler can also not be
tested on Windows for now.

* Fix skipping deployment of documentation