Make more explicit which Cabal version is supported (#2)
Previously we have only specified that the compiler has been
tested with version 2.4 of Cabal since this is the version
configured in the CI pipeline. The version 2.4 compatible
Cabal version installed on GitHub actions is version 2.4.1.0
(see https://github.com/actions/virtual-environments/blob/master/images/linux/Ubuntu1804-README.md).
Since the `new-install` command is broken in version 2.4.0.0,
I've update the README to mention the exact version of Cabal
that is officially supported.