Skip CI pipeline on pull requests that modify documentation only
Since the master branch is now protected, updates of the README
and documentation have to be performed via pull requests. In order
to avoid wasting resources, the CI pipeline will no longer run
on pull requests that modify only the README or other Markdown
files. The same rules that applied to pushes before, now apply
to pull requests as well.