Skip CI pipeline on push events #4
Since it is not possible to push to the master branch directly
and the CI pipeline runs on pull request events anyway, it is
not neccessary to run the CI pipeline on every push. If it did
it would be executed at least twice for every pull request which
would be a waste of resources.