Move Haddock deployment to own workflow
The CI pipeline does no longer run on every push but only on
the `pull_request` event. Thus, there is no documentation for
a branch until there is a pull request (However, the slug of
the generated documentation does not match the name of the
branch but is based on the pull request number).

The cleanup job that removed the deployed documentation when
the corresponding branch was deleted, no longer works since
the documentation is deployed for pull requests and not for
branches, i.e., there is nothing to remove.

Most importantly, the documentation of the master branch would
never be updated. Thus, I've removed the `deploy-docs` job from
the CI pipeline and created a separate workflow.
The new workflow runs whenever there is a push to the master
branch or a pull request is merged and Cabal configuration or
Haskell source files are modified. It builds the documentation
for the master branch and deploys the documentation to GitHub
pages.

There are no longer separate deployments for the different
branches. The workflow still deploys to `docs/master` such
that existing links are not broken and we can add documentation
for tags in the future.

I am not yet sure how this interacts with forks. Especially
if the forks disabled GitHub Actions.