Reenable CI pipeline on push events
In #4 we have decided not to run the CI pipeline on every push in
addition to the pull request runs. Otherwise the pipeline would run
up to three times for the same commit.

 1. When a member pushes to a feature branch.
 2. When that member opens a pull request.
 3. When that pull request is merged by a maintainer.

The change in #4 restricts the CI pipeline to the second run.
What we have not considered is that this prevents the pipeline
from ever running on the master branch.
That causes multiple problems in turn.

 1. The documentation for the master branch is never deployed.
    I've worked around this issue in
    cbcabdf05c57efc053e4b8dd15aa8a99062ad299
    by moving the deployment of the documentation to it's own
    GitHub actions workflow.
 2. The CI pipeline status badge in the README never updates.
 3. Most importantly, though the master has the potential to break
    even though all individual pull requests are checked.
    This can happen if the changes on the master and by the pull
    request are incompatible but cause no merge conflicts.

With this commit I reenable the CI pipeline for pushes to the master
branch. Whenever a pull request is merged, a push event is triggered
as we have found out in d8d591d3ca8a55131408095ebef1ab41eda77bab.
As a result, the CI pipeline runs twice. However, we avoid all issues
outlined above and still save one run of the pipeline.

Conceptually we can think of the first run of the pipeline as a tool
for reviewers and the user who has opened the pull request to spot
potential problems with the pull request. The second run on the
other hand is a safety net that enables the maintainer who merged
the pull request to react quickly if something breaks despite the
first run succeeded.