Avoid recompilation in CI pipeline
The cache key for `dist-newstyle` uses only the hashes of the
Cabal configuration and Cabal index. Since the modules in `src`
are updated more frequently when the Cabal configuration and
package list, modules in `src` are recompiled in every Haskell
job of the pipeline until the Cabal files change.
Now, the hash of the Haskell files in `src` is appended to the
cahce key of `dist-newstyle`. The frequent changes of Haskell
files cause the `dist-newstyle` cache to be invalidated often.
This allows the cache to be populated with newer versions of
local Haskell modules.
Unit tests and modules for the command line interface are still
expected to be recompiled frequently. Since they are not build
during the `install-dependencies` job, they never enter the
cache.