Build full package in one CI pipeline job
Separating the installation of dependencies from building the
compiler and unit tests into different jobs, did not work very
well with the population of the caches. They contained only
the cached dependencies and the internal library. The unit
tests and command line interace have been rebuilt in every
pipeline run. Now everything is build in the same job the
dependencies are installed and caches are populated.