Include `VERSION` file into `--version` output #11
The contents of the `VERSION` file that is created by the
`./tool/install.sh` script are now included into the output
of the `--version` command line option using template Haskell.
If there is not `VERSION` file (i.e., when running the compiler
using the `./tool/run.sh` script) the `--version` command falls
back to `githash`).