Use Docker for HLint CI job
The previous method of running HLint in the CI pipeline
did not work reliably, since (regardless of the version
specified in the URL), the latest HLint version was
picked. We are now using a Docker image that has the
right version of HLint preinstalled like we were doing
for the Brittany CI job already.