Ask for Git commit hash in bug report template #11 #12
Previously we asked for the output of the `freec --version`
command. The `--version` command line option includes the
compile-time hash of the Git commit. Unfortunatly, this
information is not available when `cabal new-install` is
used due to out-of-tree builds. We've explored alternatives
to add more information to the output of the `--version`
flag in #12. All of the alternatives have significant
drawbacks. Thus, I've decided to ask for the Git commit
hash in the issue template directly instead of for the
output of `--version`. This is less convinient for the user
and may cause confusion (if the installation and repository
get out of sync) but it is the best option we have at this
point in time.