Remove qualified imports from Agda's Syntax.hs #29
The module `FreeC.Backend.Agda.Syntax` imports multiple modules twice, once qualified and once unqualified. The unqualified version is needed to reexport them. This allows to use different names for the same functions (e.g. `Agda.QName` and `QName`). To avoid this we decided not to use qualified imports in this modul.