Improve error message for rigid type variables #13
If there are mutually recursive functions with type variables
that have the same name, the error message that was reported
would sometimes refer to the wrong location in source code
(the location of the last type variable that was inserted
into the environment).

Entries of rigid type variables are now inserted such that
during the unification of a type equation only those type
variables are in scope that are actually bound by the
function that caused the genration of that equation.