Avoid unnecessary monadic lifting when functions are defined as partial applications #3
Previously, the compiler performed eta abstractions any time it encountered a partial
application on a function declaration's right-hand side. These abstractions were done
by transforming the partial application into a lambda abstraction that includes the
missing arguments.

When this happens on the top-level (i.e. a function is defined as a partial application
of another function or constructor), this leads to an additional lifting into the Free
monad.
This can be avoided by applying the right-hand side to the missing arguments and
adding them to the left-hand side of the function rule explicitly.

This commit implements this optimization.