Use `check-agda.sh` script in CI pipeline
The code in the CI pipeline suffered from the same problem
as the `./tool/check-agda.sh` script did prior to
dd814820d280694ed96023af5f7ac10f22446cf1. If I'm not
mistaken, this issue should effectively have prevented the
CI pipeline from ever failing even if there was an error in
the Agda code. By reusing the `check-agda` script (which
returns an error status code if there is an error) the
problem should be fixed.