Add entries for fresh Coq identifiers
`freshCoqIdent` did not add entries to the environment.
The user could craft special identifiers that tricked the renamer
to issue the same fresh Coq identifier twice. Since all uses of
the fresh identifier are shallow (i.e., there are no user defined
binders between the introduction and use of a fresh identifier)
that's not a problem in practice. In order to avoid problems in
the future, special environment entries are now inserted when a
fresh Coq identifier is requested.