Apply formatting when outside work tree
Since we have upgraded to `actions/checkout@v2` in
1ca7a135d8d5927a1bd56819230952980e84403f
the CI job for Brittany was broken. The new version
requires Git version 2.18, but our Docker image contains
an older version (Git 2.11). Thus, the action has falled
back to the GitHub REST API. In consequence, there was no
`.git` directory. Since the `check-formatting.sh` script
skips files that are not tracked by Git, all files were
skipped. An additional check has been added to the script
to ensure that it is executed in a Git directory.
If it isn't, no files will be skipped.