`LIR.TypeCon` ow uses a snoc list #95
The old version of `LIR.TypeCon` uses a normal list to store arguments. To apply
the arguments in the right order either a `foldr` with a flipped function or
`foldl` with a reversed list has to be used. Furthermore the arguments in the
list are stored in the conceptually wrong order. This suggests that the list
should actually be a snoc list, which internally already has the right structure
and therefore exactly the needed recursion scheme.

    :          $
   / \        / \
  :   x  =>  $   x
 / \        / \
... y      ... y