Add a utility module for sharing #120
I added a module "Free.Util.Sharing" that contains
utility definitions needed for sharing.
Therefore, modules that require sharing-related
constructs can now import this module.

Currently, these modules are ND and Trace. However,
they do not import the utility module directly, but
rather import the Share effect, which exports the
utility module.

The reason for this is that it makes the most sense to me to put
the effect-specific sharing handlers  together with the effects
themselves, so they will all need to import the Share effect.

If we define an effect-generic sharing handler, the
sharing handler can be moved to the Share effect module.
In that case, ND and Trace should import the
utility module directly.