Let the handlers infer the effect stack #116
The shape and position function arguments in the
generalized effect smart constructors are now
explicit. This means that we can define an effectful
program that takes Shape and Pos as arguments
and then handle the program by letting the handlers
infer the appropriate effect stack.

The handler tests have also been adjusted so that
no fixed effect stack is used for effectful programs
like coin or traceOne.