Remove ? notation for choices #116
This commit removes the notation x ? y for Choice _ _ None x y.
The reason is that Shape and Pos should not actually be inferred
and there seems to be no pretty way of defining a ? operator that
also takes Shape and Pos.

As an alternative, the smart constructor for a choice was renamed to
Choice_ and Choice was defined as a synonym for a choice with id None.
So a non-deterministic choice between x and y can be given as
Choice Shape Pos x y.