Align arguments in type signatures #166
Unlike Brittany, Floskell does not always break `::` and the type of
the first argument into the next line if a type signature does not
fit into a single line. Thus, the first argument was not aligned with
the other arguments that were always indented by two spaces.
In order to preserve the alignment of argument comments, Floskell
is now configured such that the subsequent arguments are aligned
with the first argument.