Add ShareableArgs constraint to the type of share #149
The function share defined by the class Shareable
is now only defined for those types A that have an
associated ShareableArgs instance.
Otherwise, it is not possible to define the
call-by-need instance of Shareable, which will
make use of shareArgs.

Since ShareableArgs instances are generally defined
using cbneed, the class Shareable now requires that instances be compatible with that function, i.e.
that their associated effect stack includes the
Share effect.
This constraint is unfortunate since the call-by-name
instance doesn't technically need it as it never
makes use of shareArgs, but at present, it is
necessary.