Implement first part of Normalform instance generator #150
During the conversion of a data declaration, a Normalform instance
will be generated.
This commit implements a part of that generator that generates
most of the nf' function. The resulting code is not yet
valid because the type signature is still missing, but if the
type signature is added manually, the generated code is valid.

We are able to consider types with nested recursion and mutually
recursive types, but I have not considered type synonyms yet.
That will also be added later.