Apply share operator explicitly in examples #149
In the examples in the sharing handler tests, the
share operator is now applied explicitly. This way,
we can make sure that the appropriate arguments
(especially type class instances) are used.

To do that, the Shareable instance given to the
example functions no longer features implicit generalization
and both the Share and the ShareableArgs constraint are given as
(implicit) arguments to the function so they can be used in
the explicit application of share.