Use a function instead of ShareableArgs instance in cbneed #149
Instead of taking a ShareableArgs instance as an argument and
using its shareArgs function to share components of data structures,
the cbneed operator now takes a function shrArgs : A -> Free Shape Pos A
and uses that.
This makes it easier to define the function shareArgs for recursive
data types as cbneed can be called with the function being defined
as an argument. This prevents having to inline the definition of cbneed.