Add empty sharing handler #177
This commits adds a sharing handler to Share for effects that do
not require sharing (as of the time of this commit, that is
every effect except for Trace and ND).

Previously, to handle a program that contains sharing, but no
relevant effect, one of the relevant effects would unnecessarily
be added to the effect stack.
With this handler, this is no longer necessary.

However, this handler should NOT be used to handle relevant effects,
as no sharing identifiers will actually be assigned.