Implement Handler instances #177
Handler instances for most combinations of effects have been
defined in a new module Free.HandlerInstances. This way, the
Handler class can be imported without importing all instances.

Instances exist for most combinations of effects in a fixed order.
The only combinations missing are those that include all three
effects sharing, non-determinism and tracing.
This is because currently, we can model sharing for at most one
effect in a program.