Use fresh variables for unification and adjust local environments #150
Before the return type of a data constructor is unified with a type
expression, all variables (underscores) in the type expression are
replaced with fresh variables to prevent unification failures.

Additionally, the naming of the instance and top-level functions is
now done outside of a local environment so that those names are
registered globally and no name clashes can occur.
Local functions and variables are still named inside a local environment.