Change handling order of non-deterministic partial programs #183
In programs containing both the Maybe effect and the non-determinism
effect, Maybe is now handled after non-determinism. This means that,
if an error occurs in one non-deterministic branch, the entire
computation will throw an error.
Previously, only the branch with the error was affected.
However, error "Error" 1 ? 1 evaluates to an error in Curry, not
the choice between an error and 1.

The handling order for tracing in combination with maybe remains the
same, however, as trace "Log" (error "Error") still logs the tracing
message in Haskell (and Curry) before throwing an error.