Add utility function to map subterms #158
We often have to traverse the AST to apply a function on some
nodes but want to leave all other nodes unmodified. The existing
subterm functions can be used to traverse the AST without
having to handle every constructor explicitly. The new helper
function provides a convenient interface for applying a
function on a node and all of its subterms.