Remove second type argument from Normalform class #202
The second type argument to Normalform represents the result
type of the normalization. That result type is the same as the
input type A, but with Identity.Shape and Identity.Pos replacing
all occurrences of Shape and Pos in A.

In its place, a constant nType has been added to the class that
represents that same result type.

nType does not have to be explicitly implemented when defining a
Normalform instance as it can be inferred from the return type
of nf'.