Add Handler class with two functions #202
The Handler class has been added again, but with only one
type parameter.
Instead of the result type parameter B, the Handler class now
contains a function transformType : Type -> Type that describes
how an input type changes when it is handled.
The result type of the handler is now determined by applying
transformType to the result type of normalizing the input type A.