Change type of strategy argument #202
The type of the strategy argument was changed from
Strategy Shape Pos to
forall (Shape : Type) (Pos : Shape -> Type)
       `{Injectable Share.Shape Share.Pos Shape Pos),
  Strategy Shape Pos.

This makes it easier to use in theorems because it is possible
to write it simply as "Cbv", "Cbn" or "Cbneed" instead of specifying
Shape and Pos.

A notation "EvaluationStrategy" for the above type has been added and
is used by the compiler.