Add handling to quickCheck #202
Property is now no longer a synonym for Prop, but a function
that takes a handler and returns a Prop.
Because the handler is not part of the property itself, there
are now two different functions for creating properties:
quickCheck quantifies the handler and quickCheckHandle takes
a handler as an additional argument and applies it.

To use quickCheckHandle on a property, the Shape and Pos must be
inferred, so it is necessary to write underscores in their place
(or the exact effect stack that the handler uses).

quickCheck can be used like before, but to choose an evaluation
strategy (which is recommended), one should write
quickCheck (fun Shape Pos I => prop_x Shape Pos I <Cb*>), where
prop_x is a property and <Cb*> an evaluation strategy.