Generate helper functions for `case` of aliases #209
The termination checker was previously extended to allow
aliases of the decreasing argument or structurally smaller
variables to be created using `let`-expressions without
affecting termination. However, the Coq backend has not
been extended accordingly. The missing helper functions
are generated now but the arguments of the helper functions
are not passed correctly.