Inline lambda abstractions bound by `let`-expressions #169
Coq currently does not accept the `mapRose` example because it
does not know how the lambda abstraction (that is introduced by
η-conversion) is invoked. Coq can recognize that the function
terminates if we remove the corresponding `call` operations (i.e.,
`let`-bindings in IR). This does not change the semantics of the
program since lambda abstractions are not evaluated until they are
invoked in every evaluation strategy.