Remove `call` operator #169
The `call` operator was an optimization to avoid the generation of
unnecessary `share`s which would complicate proofs in a call-by-need
setting. However, even if we transform expressions into a flat form
additional `share` operators are needed to model sharing correctly
if we are using the `call` operator. Since we only want to generate
the flat form, we cannot apply the optimization or have to find a
way to model sharing that also works in a call-by-value setting.